Not crashing when going outside a map in a direction that don´t have a new map
