+++ src/main/java/model/MapStates/Markena.java
@@ -17,9 +17,9 @@ public enum Markena implements MapState{
                 return Karhuset.KARHUSET;
             case LEFT:
                 return Maskin.MASKIN;
-
+            default:
+                return Markena.MARKENA;
         }
-        return null;
     }
 
     @Override
