+++ src/main/java/model/MapStates/Maskin.java
@@ -17,9 +17,10 @@ public enum Maskin implements MapState{
                 return Markena.MARKENA;
             case UP:
                 return Sandladan.SANDLADAN;
+            default:
+                return Maskin.MASKIN;
 
         }
-        return null;
     }
 
     @Override
